Synchronize code for oneplus LE2101_13.1.0.581(EX01)

+++ drivers/misc/colorctrl/colorctrl.c
@@ -861,7 +861,7 @@ static ssize_t proc_colorctrl_read(struct file *file, char __user *user_buf, siz
 
     colorctrl_set_awake(cd, true);
     mutex_lock(&cd->rw_lock);
-    snprintf(page, PAGESIZE - 1, "%u\n", cd->color_status);
+    snprintf(page, PAGESIZE - 1, "%d\n", (int)cd->color_status);
     ret = simple_read_from_buffer(user_buf, count, ppos, page, strlen(page));
     mutex_unlock(&cd->rw_lock);
     COLOR_INFO("read value: %d.", cd->color_status);
@@ -897,7 +897,7 @@ static ssize_t proc_temperature_control_read(struct file *file, char __user *use
         snprintf(page, PAGESIZE - 1, "failed to get fitting temperature\n");
     } else {
         COLOR_INFO("current fitting temperature is : %d", temp);
-        snprintf(page, PAGESIZE - 1, "%d\n", temp);
+        snprintf(page, PAGESIZE - 1, "%lld\n", temp);
     }
 
     mutex_unlock(&cd->rw_lock);
